backend/requirements.txt
matplotlib
pandas
numpy
Flask==2.0.3
mysql-connector==2.2.9
